// src/lib/kubb/types/Message.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type Message = {
  /**
   * @minLength 0
   * @type integer | undefined
   */
  chat_id?: number;
  /**
   * @type string | undefined
   */
  classification_result?: string;
  /**
   * @type string | undefined
   */
  content?: string;
  /**
   * @minLength 0
   * @type integer
   */
  id?: number | null;
  /**
   * @type string | undefined
   */
  image?: string;
  /**
   * @type string | undefined
   */
  image_to_classify?: string;
  /**
   * @type boolean | undefined
   */
  is_typing?: boolean;
  /**
   * @type string | undefined
   */
  iv?: string;
  /**
   * @minLength 0
   * @type integer | undefined
   */
  receiver_id?: number;
  /**
   * @minLength 0
   * @type integer | undefined
   */
  sender_id?: number;
  /**
   * @type string | undefined
   */
  sender_name?: string;
  /**
   * @type string | undefined, date-time
   */
  timestamp?: string;
  /**
   * @type string | undefined
   */
  type?: string;
};